:sparkles: 【需求变更】供水温度上下限变成浮点型

+++ ems-module-power/ems-module-power-biz/src/main/java/cn/bitlinks/ems/module/power/controller/admin/report/supplywatertmp/vo/SupplyWaterTmpSettingsPageReqVO.java
@@ -6,6 +6,8 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
 
+import java.math.BigDecimal;
+
 /**
  * @author liumingqiang
  */
@@ -32,8 +34,8 @@ public class SupplyWaterTmpSettingsPageReqVO extends PageParam {
     private String energyParamName;
 
     @Schema(description = "上限")
-    private Integer max;
+    private BigDecimal max;
 
     @Schema(description = "下限")
-    private Integer min;
+    private BigDecimal min;
 }
