+++ ems-module-power/ems-module-power-biz/src/main/java/cn/bitlinks/ems/module/power/controller/admin/report/supplywatertmp/vo/SupplyWaterTmpSettingsRespVO.java
@@ -5,6 +5,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
 import javax.validation.constraints.NotEmpty;
+import java.math.BigDecimal;
 
 /**
  * @author liumingqiang
@@ -31,10 +32,10 @@ public class SupplyWaterTmpSettingsRespVO {
     private String energyParamName;
 
     @Schema(description = "上限")
-    private Integer max;
+    private BigDecimal max;
 
     @Schema(description = "下限")
-    private Integer min;
+    private BigDecimal min;
 
 
 }
