+++ ems-module-power/ems-module-power-biz/src/main/java/cn/bitlinks/ems/module/power/dal/dataobject/report/supplywatertmp/SupplyWaterTmpSettingsDO.java
@@ -7,6 +7,8 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 /**
  * 供应分设置析 DO
  *
@@ -50,9 +52,9 @@ public class SupplyWaterTmpSettingsDO extends BaseDO {
     /**
      * 上限
      */
-    private Integer max;
+    private BigDecimal max;
     /**
      * 下限
      */
-    private Integer min;
+    private BigDecimal min;
 }
